+++ bounties_api/std_bounties/serializers.py
@@ -1,7 +1,7 @@
 from django.apps import apps
 from rest_framework import serializers
 from bounties.serializers import CreatableSlugRelatedField
-from std_bounties.models import Bounty, Fulfillment, Category, RankedCategory, Token, DraftBounty, Comment, Review, Language
+from std_bounties.models import Bounty, Fulfillment, Category, RankedCategory, Token, DraftBounty, Comment, Review
 from std_bounties.client_helpers import map_token_data
 from std_bounties.constants import STAGE_CHOICES
 from django.db import transaction
@@ -34,13 +34,6 @@ class CategorySerializer(serializers.ModelSerializer):
         model = Category
         fields = '__all__'
 
-class LanguageSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = Language
-        fields = [
-            'name',
-            'normalized_name'
-        ]
 
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
