+++ bounties_api/std_bounties/urls.py
@@ -9,7 +9,6 @@ router.register(r'^bounty/draft', views.DraftBountyWriteViewSet)
 router.register(r'^bounty', views.BountyViewSet)
 router.register(r'^fulfillment', views.FulfillmentViewSet)
 router.register(r'^category', views.CategoryViewSet)
-router.register(r'^languages', views.LanguageViewSet)
 
 urlpatterns = [
     url(r'^leaderboard/issuer/$', views.LeaderboardIssuer.as_view()),
