+++ bounties_api/std_bounties/views.py
@@ -10,9 +10,9 @@ from rest_framework.views import APIView
 from bounties.utils import dictfetchall, extractInParams, sqlGenerateOrList, limitOffsetParams
 from std_bounties.constants import STAGE_CHOICES
 from std_bounties.queries import LEADERBOARD_ISSUER_QUERY, LEADERBOARD_FULFILLER_QUERY
-from std_bounties.serializers import BountySerializer, FulfillmentSerializer, RankedCategorySerializer, LanguageSerializer, LeaderboardIssuerSerializer, LeaderboardFulfillerSerializer, TokenSerializer, DraftBountyWriteSerializer, CommentSerializer, ReviewSerializer
-from std_bounties.models import Bounty, DraftBounty, Fulfillment, RankedCategory, Token, Comment, Language
-from std_bounties.filters import BountiesFilter, FulfillmentsFilter, LanguageFilter, RankedCategoryFilter
+from std_bounties.serializers import BountySerializer, FulfillmentSerializer, RankedCategorySerializer, LeaderboardIssuerSerializer, LeaderboardFulfillerSerializer, TokenSerializer, DraftBountyWriteSerializer, CommentSerializer, ReviewSerializer
+from std_bounties.models import Bounty, DraftBounty, Fulfillment, RankedCategory, Token, Comment
+from std_bounties.filters import BountiesFilter, FulfillmentsFilter, RankedCategoryFilter
 from user.permissions import AuthenticationPermission, UserObjectPermissions
 from notifications.notification_client import NotificationClient
 from rest_framework.filters import OrderingFilter, SearchFilter
@@ -140,11 +140,6 @@ class CategoryViewSet(viewsets.ReadOnlyModelViewSet):
     ordering_fields = ('total_count',)
     ordering = ('-total_count',)
     search_fields = ('normalized_name',)
-
-
-class LanguageViewSet(viewsets.ReadOnlyModelViewSet):
-    serializer_class = LanguageSerializer
-    queryset = Language.objects.order_by('name')
 
 
 class UserProfile(APIView):
